+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `entrait_export` macro and `export` option, for exporting optional mocks from libraries.
 - `async-trait` feature for adding a re-export of the async-trait crate.
 - `use-async-trait` and `use-associated-future` features for global selection of async strategy.
+- Support for generic functions.
 ### Changed
 - Restructure lib docs a bit.
 ### Removed